mate-check-update.py: fix stripping of PKGREVISION values
Strip PKGREVISION strings from the variable that represents the pkgsrc
version number, not the upstream version.

Before:

mate-panel , 1.24.0nb2 , 1.24.1 , no

After:

mate-panel , 1.24.0 , 1.24.1 , yes

(The meanings of the variables "lv" and "uv" are reversed versus
the same names in xfce4-check-update.py. I kept their use internally
consistent here, to minimize the diff in this change. But it is
potentially confusing.)